+++ dart/lib/src/sentry_item_type.dart
@@ -4,5 +4,6 @@ class SentryItemType {
   static const String attachment = 'attachment';
   static const String transaction = 'transaction';
   static const String clientReport = 'client_report';
+  static const String profile = 'profile';
   static const String unknown = '__unknown__';
 }